Verbose numbers and Sesame Street reference
A follower request, I've added the classic laugh from the count on
Sesame Street to the tweets. I've also spent probably too long making it
so that the numbers are spelled out into words, as you might expect the
count to do. I'm sure there are lots of mistakes, but I have work in
four hours, and this is a gimmick project, so I don't care.