Modules exist
Someone pointed out that num2words
(https://github.com/savoirfairelinux/num2words) exists and that I should
use it, which I am super grateful for because my version was not very
Raven. The replace is there because "and" in verbose numbers indicates
decimal places, and this is a very basic fact that is ignored always for
reasons unbeknownst to me.